+++ airbyte-commons-server/src/test/java/io/airbyte/commons/server/handlers/AttemptHandlerTest.java
@@ -202,7 +202,7 @@ class AttemptHandlerTest {
     Mockito.when(path.resolve(Mockito.anyString()))
         .thenReturn(path);
 
-    final Path expectedRoot = TemporalUtils.getJobRoot(path, String.valueOf(JOB_ID), ATTEMPT_NUMBER);
+    final Path expectedRoot = TemporalUtils.getJobRoot(path, String.valueOf(JOB_ID), (long) ATTEMPT_NUMBER);
     final Path expectedLogPath = expectedRoot.resolve(LogClientSingleton.LOG_FILENAME);
 
     Mockito.when(jobPersistence.createAttempt(JOB_ID, expectedLogPath))
